requirements.txt
Django==5.0.2
djangorestframework==3.14.0
elasticsearch==8.12.0
djangorestframework-simplejwt==5.3.0
django-elasticsearch-dsl==8.0